--- cpp/mqtt_tests/helpers/prepare_test_services.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "apigear/mqtt/mqttservice.h"
-
-void prepareStaticServices(std::shared_ptr<ApiGear::MQTT::Service> serviceAdapter);
