--- cpp/mqtt_tests/helpers/test_message_topics.h
@@ -1,9 +1,0 @@
-#pragma once
-#include <string>
-
-// Messages to be send from client(s) to service adapter to inform that it started test or stopped test
-namespace TestMessages
-{
-const std::string clientStartsTestTopic = "StartTestTopic";
-const std::string clientStopsTestTopic = "StopTestTopic";
-}
