+++ cpp/mqtt_tests/many_objects/CMakeLists.txt
@@ -13,7 +13,7 @@ set( CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin )
 add_executable(mqtt_multi_client_setIntProperty "client_setIntProperty.cpp")
 add_executable(mqtt_multi_client_setFloatProperty "client_setFloatProperty.cpp")
 add_executable(mqtt_multi_client_setStringProperty "client_setStringProperty.cpp")
-add_executable(mqtt_multi_service "server.cpp" "../helpers/prepare_test_services.cpp" )
+add_executable(mqtt_multi_service "server.cpp" "helpers/prepare_test_services.cpp" )
 
 target_link_libraries(mqtt_multi_client_setIntProperty api::api-mqtt)
 target_link_libraries(mqtt_multi_client_setFloatProperty api::api-mqtt)
