+++ cpp/mqtt_tests/many_objects/helpers/prepare_test_services.cpp
@@ -1,0 +1,343 @@
+#include "prepare_test_services.h"
+
+#include "api/generated/mqtt/testapi0service.h"
+#include "api/generated/mqtt/testapi1service.h"
+#include "api/generated/mqtt/testapi2service.h"
+#include "api/generated/mqtt/testapi3service.h"
+#include "api/generated/mqtt/testapi4service.h"
+#include "api/generated/mqtt/testapi5service.h"
+#include "api/generated/mqtt/testapi6service.h"
+#include "api/generated/mqtt/testapi7service.h"
+#include "api/generated/mqtt/testapi8service.h"
+#include "api/generated/mqtt/testapi9service.h"
+
+#include "api/generated/mqtt/testapi10service.h"
+#include "api/generated/mqtt/testapi11service.h"
+#include "api/generated/mqtt/testapi12service.h"
+#include "api/generated/mqtt/testapi13service.h"
+#include "api/generated/mqtt/testapi14service.h"
+#include "api/generated/mqtt/testapi15service.h"
+#include "api/generated/mqtt/testapi16service.h"
+#include "api/generated/mqtt/testapi17service.h"
+#include "api/generated/mqtt/testapi18service.h"
+#include "api/generated/mqtt/testapi19service.h"
+
+#include "api/generated/mqtt/testapi20service.h"
+#include "api/generated/mqtt/testapi21service.h"
+#include "api/generated/mqtt/testapi22service.h"
+#include "api/generated/mqtt/testapi23service.h"
+#include "api/generated/mqtt/testapi24service.h"
+#include "api/generated/mqtt/testapi25service.h"
+#include "api/generated/mqtt/testapi26service.h"
+#include "api/generated/mqtt/testapi27service.h"
+#include "api/generated/mqtt/testapi28service.h"
+#include "api/generated/mqtt/testapi29service.h"
+
+#include "api/generated/mqtt/testapi30service.h"
+#include "api/generated/mqtt/testapi31service.h"
+#include "api/generated/mqtt/testapi32service.h"
+#include "api/generated/mqtt/testapi33service.h"
+#include "api/generated/mqtt/testapi34service.h"
+#include "api/generated/mqtt/testapi35service.h"
+#include "api/generated/mqtt/testapi36service.h"
+#include "api/generated/mqtt/testapi37service.h"
+#include "api/generated/mqtt/testapi38service.h"
+#include "api/generated/mqtt/testapi39service.h"
+
+#include "api/generated/mqtt/testapi40service.h"
+#include "api/generated/mqtt/testapi41service.h"
+#include "api/generated/mqtt/testapi42service.h"
+#include "api/generated/mqtt/testapi43service.h"
+#include "api/generated/mqtt/testapi44service.h"
+#include "api/generated/mqtt/testapi45service.h"
+#include "api/generated/mqtt/testapi46service.h"
+#include "api/generated/mqtt/testapi47service.h"
+#include "api/generated/mqtt/testapi48service.h"
+#include "api/generated/mqtt/testapi49service.h"
+
+#include "api/generated/mqtt/testapi50service.h"
+#include "api/generated/mqtt/testapi51service.h"
+#include "api/generated/mqtt/testapi52service.h"
+#include "api/generated/mqtt/testapi53service.h"
+#include "api/generated/mqtt/testapi54service.h"
+#include "api/generated/mqtt/testapi55service.h"
+#include "api/generated/mqtt/testapi56service.h"
+#include "api/generated/mqtt/testapi57service.h"
+#include "api/generated/mqtt/testapi58service.h"
+#include "api/generated/mqtt/testapi59service.h"
+
+#include "api/generated/mqtt/testapi60service.h"
+#include "api/generated/mqtt/testapi61service.h"
+#include "api/generated/mqtt/testapi62service.h"
+#include "api/generated/mqtt/testapi63service.h"
+#include "api/generated/mqtt/testapi64service.h"
+#include "api/generated/mqtt/testapi65service.h"
+#include "api/generated/mqtt/testapi66service.h"
+#include "api/generated/mqtt/testapi67service.h"
+#include "api/generated/mqtt/testapi68service.h"
+#include "api/generated/mqtt/testapi69service.h"
+
+#include "api/generated/mqtt/testapi70service.h"
+#include "api/generated/mqtt/testapi71service.h"
+#include "api/generated/mqtt/testapi72service.h"
+#include "api/generated/mqtt/testapi73service.h"
+#include "api/generated/mqtt/testapi74service.h"
+#include "api/generated/mqtt/testapi75service.h"
+#include "api/generated/mqtt/testapi76service.h"
+#include "api/generated/mqtt/testapi77service.h"
+#include "api/generated/mqtt/testapi78service.h"
+#include "api/generated/mqtt/testapi79service.h"
+
+#include "api/generated/mqtt/testapi80service.h"
+#include "api/generated/mqtt/testapi81service.h"
+#include "api/generated/mqtt/testapi82service.h"
+#include "api/generated/mqtt/testapi83service.h"
+#include "api/generated/mqtt/testapi84service.h"
+#include "api/generated/mqtt/testapi85service.h"
+#include "api/generated/mqtt/testapi86service.h"
+#include "api/generated/mqtt/testapi87service.h"
+#include "api/generated/mqtt/testapi88service.h"
+#include "api/generated/mqtt/testapi89service.h"
+
+#include "api/generated/mqtt/testapi90service.h"
+#include "api/generated/mqtt/testapi91service.h"
+#include "api/generated/mqtt/testapi92service.h"
+#include "api/generated/mqtt/testapi93service.h"
+#include "api/generated/mqtt/testapi94service.h"
+#include "api/generated/mqtt/testapi95service.h"
+#include "api/generated/mqtt/testapi96service.h"
+#include "api/generated/mqtt/testapi97service.h"
+#include "api/generated/mqtt/testapi98service.h"
+#include "api/generated/mqtt/testapi99service.h"
+
+#include "api/implementation/testapi0.h"
+#include "api/implementation/testapi1.h"
+#include "api/implementation/testapi2.h"
+#include "api/implementation/testapi3.h"
+#include "api/implementation/testapi4.h"
+#include "api/implementation/testapi5.h"
+#include "api/implementation/testapi6.h"
+#include "api/implementation/testapi7.h"
+#include "api/implementation/testapi8.h"
+#include "api/implementation/testapi9.h"
+
+#include "api/implementation/testapi10.h"
+#include "api/implementation/testapi11.h"
+#include "api/implementation/testapi12.h"
+#include "api/implementation/testapi13.h"
+#include "api/implementation/testapi14.h"
+#include "api/implementation/testapi15.h"
+#include "api/implementation/testapi16.h"
+#include "api/implementation/testapi17.h"
+#include "api/implementation/testapi18.h"
+#include "api/implementation/testapi19.h"
+
+#include "api/implementation/testapi20.h"
+#include "api/implementation/testapi21.h"
+#include "api/implementation/testapi22.h"
+#include "api/implementation/testapi23.h"
+#include "api/implementation/testapi24.h"
+#include "api/implementation/testapi25.h"
+#include "api/implementation/testapi26.h"
+#include "api/implementation/testapi27.h"
+#include "api/implementation/testapi28.h"
+#include "api/implementation/testapi29.h"
+
+#include "api/implementation/testapi30.h"
+#include "api/implementation/testapi31.h"
+#include "api/implementation/testapi32.h"
+#include "api/implementation/testapi33.h"
+#include "api/implementation/testapi34.h"
+#include "api/implementation/testapi35.h"
+#include "api/implementation/testapi36.h"
+#include "api/implementation/testapi37.h"
+#include "api/implementation/testapi38.h"
+#include "api/implementation/testapi39.h"
+
+#include "api/implementation/testapi40.h"
+#include "api/implementation/testapi41.h"
+#include "api/implementation/testapi42.h"
+#include "api/implementation/testapi43.h"
+#include "api/implementation/testapi44.h"
+#include "api/implementation/testapi45.h"
+#include "api/implementation/testapi46.h"
+#include "api/implementation/testapi47.h"
+#include "api/implementation/testapi48.h"
+#include "api/implementation/testapi49.h"
+
+#include "api/implementation/testapi50.h"
+#include "api/implementation/testapi51.h"
+#include "api/implementation/testapi52.h"
+#include "api/implementation/testapi53.h"
+#include "api/implementation/testapi54.h"
+#include "api/implementation/testapi55.h"
+#include "api/implementation/testapi56.h"
+#include "api/implementation/testapi57.h"
+#include "api/implementation/testapi58.h"
+#include "api/implementation/testapi59.h"
+
+#include "api/implementation/testapi60.h"
+#include "api/implementation/testapi61.h"
+#include "api/implementation/testapi62.h"
+#include "api/implementation/testapi63.h"
+#include "api/implementation/testapi64.h"
+#include "api/implementation/testapi65.h"
+#include "api/implementation/testapi66.h"
+#include "api/implementation/testapi67.h"
+#include "api/implementation/testapi68.h"
+#include "api/implementation/testapi69.h"
+
+#include "api/implementation/testapi70.h"
+#include "api/implementation/testapi71.h"
+#include "api/implementation/testapi72.h"
+#include "api/implementation/testapi73.h"
+#include "api/implementation/testapi74.h"
+#include "api/implementation/testapi75.h"
+#include "api/implementation/testapi76.h"
+#include "api/implementation/testapi77.h"
+#include "api/implementation/testapi78.h"
+#include "api/implementation/testapi79.h"
+
+#include "api/implementation/testapi80.h"
+#include "api/implementation/testapi81.h"
+#include "api/implementation/testapi82.h"
+#include "api/implementation/testapi83.h"
+#include "api/implementation/testapi84.h"
+#include "api/implementation/testapi85.h"
+#include "api/implementation/testapi86.h"
+#include "api/implementation/testapi87.h"
+#include "api/implementation/testapi88.h"
+#include "api/implementation/testapi89.h"
+
+#include "api/implementation/testapi90.h"
+#include "api/implementation/testapi91.h"
+#include "api/implementation/testapi92.h"
+#include "api/implementation/testapi93.h"
+#include "api/implementation/testapi94.h"
+#include "api/implementation/testapi95.h"
+#include "api/implementation/testapi96.h"
+#include "api/implementation/testapi97.h"
+#include "api/implementation/testapi98.h"
+#include "api/implementation/testapi99.h"
+
+
+template< class MqttServiceAdapter, class InterfaceService>
+void addService(std::shared_ptr<ApiGear::MQTT::Service> serviceAdapter)
+{
+    std::shared_ptr<InterfaceService> service = std::make_shared<InterfaceService>();
+    static auto mqttServiceAdapter = std::make_shared<MqttServiceAdapter>(service, serviceAdapter);
+}
+
+
+void prepareStaticServices(std::shared_ptr<ApiGear::MQTT::Service> serviceAdapter)
+{
+    addService <Cpp::Api::MQTT::TestApi0Service, Cpp::Api::TestApi0>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi1Service, Cpp::Api::TestApi1>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi2Service, Cpp::Api::TestApi2>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi3Service, Cpp::Api::TestApi3>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi4Service, Cpp::Api::TestApi4>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi5Service, Cpp::Api::TestApi5>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi6Service, Cpp::Api::TestApi6>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi7Service, Cpp::Api::TestApi7>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi8Service, Cpp::Api::TestApi8>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi9Service, Cpp::Api::TestApi9>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi10Service, Cpp::Api::TestApi10>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi11Service, Cpp::Api::TestApi11>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi12Service, Cpp::Api::TestApi12>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi13Service, Cpp::Api::TestApi13>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi14Service, Cpp::Api::TestApi14>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi15Service, Cpp::Api::TestApi15>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi16Service, Cpp::Api::TestApi16>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi17Service, Cpp::Api::TestApi17>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi18Service, Cpp::Api::TestApi18>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi19Service, Cpp::Api::TestApi19>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi20Service, Cpp::Api::TestApi20>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi21Service, Cpp::Api::TestApi21>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi22Service, Cpp::Api::TestApi22>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi23Service, Cpp::Api::TestApi23>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi24Service, Cpp::Api::TestApi24>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi25Service, Cpp::Api::TestApi25>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi26Service, Cpp::Api::TestApi26>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi27Service, Cpp::Api::TestApi27>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi28Service, Cpp::Api::TestApi28>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi29Service, Cpp::Api::TestApi29>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi30Service, Cpp::Api::TestApi30>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi31Service, Cpp::Api::TestApi31>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi32Service, Cpp::Api::TestApi32>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi33Service, Cpp::Api::TestApi33>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi34Service, Cpp::Api::TestApi34>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi35Service, Cpp::Api::TestApi35>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi36Service, Cpp::Api::TestApi36>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi37Service, Cpp::Api::TestApi37>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi38Service, Cpp::Api::TestApi38>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi39Service, Cpp::Api::TestApi39>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi40Service, Cpp::Api::TestApi40>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi41Service, Cpp::Api::TestApi41>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi42Service, Cpp::Api::TestApi42>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi43Service, Cpp::Api::TestApi43>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi44Service, Cpp::Api::TestApi44>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi45Service, Cpp::Api::TestApi45>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi46Service, Cpp::Api::TestApi46>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi47Service, Cpp::Api::TestApi47>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi48Service, Cpp::Api::TestApi48>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi49Service, Cpp::Api::TestApi49>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi50Service, Cpp::Api::TestApi50>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi51Service, Cpp::Api::TestApi51>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi52Service, Cpp::Api::TestApi52>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi53Service, Cpp::Api::TestApi53>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi54Service, Cpp::Api::TestApi54>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi55Service, Cpp::Api::TestApi55>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi56Service, Cpp::Api::TestApi56>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi57Service, Cpp::Api::TestApi57>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi58Service, Cpp::Api::TestApi58>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi59Service, Cpp::Api::TestApi59>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi60Service, Cpp::Api::TestApi60>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi61Service, Cpp::Api::TestApi61>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi62Service, Cpp::Api::TestApi62>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi63Service, Cpp::Api::TestApi63>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi64Service, Cpp::Api::TestApi64>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi65Service, Cpp::Api::TestApi65>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi66Service, Cpp::Api::TestApi66>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi67Service, Cpp::Api::TestApi67>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi68Service, Cpp::Api::TestApi68>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi69Service, Cpp::Api::TestApi69>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi70Service, Cpp::Api::TestApi70>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi71Service, Cpp::Api::TestApi71>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi72Service, Cpp::Api::TestApi72>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi73Service, Cpp::Api::TestApi73>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi74Service, Cpp::Api::TestApi74>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi75Service, Cpp::Api::TestApi75>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi76Service, Cpp::Api::TestApi76>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi77Service, Cpp::Api::TestApi77>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi78Service, Cpp::Api::TestApi78>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi79Service, Cpp::Api::TestApi79>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi80Service, Cpp::Api::TestApi80>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi81Service, Cpp::Api::TestApi81>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi82Service, Cpp::Api::TestApi82>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi83Service, Cpp::Api::TestApi83>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi84Service, Cpp::Api::TestApi84>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi85Service, Cpp::Api::TestApi85>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi86Service, Cpp::Api::TestApi86>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi87Service, Cpp::Api::TestApi87>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi88Service, Cpp::Api::TestApi88>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi89Service, Cpp::Api::TestApi89>(serviceAdapter);
+
+    addService <Cpp::Api::MQTT::TestApi90Service, Cpp::Api::TestApi90>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi91Service, Cpp::Api::TestApi91>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi92Service, Cpp::Api::TestApi92>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi93Service, Cpp::Api::TestApi93>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi94Service, Cpp::Api::TestApi94>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi95Service, Cpp::Api::TestApi95>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi96Service, Cpp::Api::TestApi96>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi97Service, Cpp::Api::TestApi97>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi98Service, Cpp::Api::TestApi98>(serviceAdapter);
+    addService <Cpp::Api::MQTT::TestApi99Service, Cpp::Api::TestApi99>(serviceAdapter);
+}
