+++ cpp/mqtt_tests/many_objects/helpers/prepare_test_services.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "apigear/mqtt/mqttservice.h"
+
+void prepareStaticServices(std::shared_ptr<ApiGear::MQTT::Service> serviceAdapter);
