+++ cpp/nats_tests/nats_network_protocol_handler_for_test.hpp
