+++ cpp/nats_tests/single_object_many_threads/client_syncSetIntProperty.cpp
@@ -1,7 +1,7 @@
 
 #include "api/generated/nats/testapi0client.h"
 
-#include "../helpers/nats_network_protocol_handler_for_test.hpp"
+#include "../nats_network_protocol_handler_for_test.hpp"
 #include "../../latency_helpers/latency_helpers.h"
 #include "../../scenario_templates/single_object_many_threads/executeTestFunction.h"
