+++ cpp/olink_tests/olink_core_only_tests/olink_core_no_network_single_object_asyncIntMethod.cpp
@@ -9,9 +9,9 @@
 #include "api/implementation//testapi0.h"
 
 #include "../../scenario_templates/single_object_many_threads/executeTestFunction.h"
+#include "../../latency_helpers/latency_helpers.h"
 #include "../helpers/inspected_sink.h"
-#include "../helpers/latency_helpers.h"
-#include "../helpers/olink_handler_no_network.hpp"
+#include "../olink_handler_no_network.hpp"
 
 #include <memory>
 #include <chrono>
