+++ cpp/olink_tests/olink_core_only_tests/olink_core_no_network_single_object_setIntProperty.cpp
@@ -8,10 +8,10 @@
 #include "api/generated/olink/testapi0service.h"
 #include "api/implementation//testapi0.h"
 
-#include "../helpers/inspected_sink.h"
-#include "../../latency_helpers/latency_helpers.h"
-#include "../helpers/olink_handler_no_network.hpp"
 #include "../../scenario_templates/single_object_many_threads/executeTestFunction.h"
+#include "../../latency_helpers/latency_helpers.h"
+#include "../olink_handler_no_network.hpp"
+#include "../helpers/inspected_sink.h"
 
 #include <memory>
 #include <chrono>
