+++ cpp/olink_tests/olink_core_only_tests/olink_core_no_network_single_object_syncIntMethodCall.cpp
@@ -11,7 +11,7 @@
 #include "../../scenario_templates/single_object_many_threads/executeTestFunction.h"
 #include "../helpers/inspected_sink.h"
 #include "../../latency_helpers/latency_helpers.h"
-#include "../helpers/olink_handler_no_network.hpp"
+#include "../olink_handler_no_network.hpp"
 
 #include <memory>
 #include <chrono>
