+++ cpp/olink_tests/single_object_many_threads/client_syncIntMethodCall.cpp
@@ -1,7 +1,7 @@
 #include "../helpers/inspected_sink.h"
 #include "api/generated/olink/testapi0client.h"
 
-#include "../helpers/olink_network_protocol_handler_for_test.hpp"
+#include "../olink_network_protocol_handler_for_test.hpp"
 #include "../../latency_helpers/latency_helpers.h"
 #include "../../scenario_templates/single_object_many_threads/executeTestFunction.h"
